getCDAPConfig and getHadoopConfig in MetaClient
